+++ user/subsystem/rs485.h
@@ -12,6 +12,6 @@ typedef struct {
 
 void init_rs485(RS485Port *);
 void rs485_transmit(RS485Port *, uint8_t * buf, uint16_t size);
-bool_t rs485_receive(RS485Port *, uint16_t * buf, uint16_t size, uint32_t timeout);
+bool_t rs485_receive(RS485Port *, uint8_t * buf, uint16_t size, uint32_t timeout);
 
 #endif
